Fixing Namespace, renaming ViewModelBase => DialogViewModelBase
